split shard.go into separate files per lifecyle step
