+++ adapters/repos/db/shard.go
@@ -15,19 +15,15 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"os"
 	"path"
-	"runtime/debug"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	enterrors "github.com/weaviate/weaviate/entities/errors"
-	entsentry "github.com/weaviate/weaviate/entities/sentry"
 
 	"github.com/go-openapi/strfmt"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"github.com/weaviate/weaviate/adapters/repos/db/helpers"
 	"github.com/weaviate/weaviate/adapters/repos/db/indexcheckpoint"
 	"github.com/weaviate/weaviate/adapters/repos/db/indexcounter"
@@ -35,15 +31,9 @@ import (
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv"
 	"github.com/weaviate/weaviate/adapters/repos/db/propertyspecific"
 	"github.com/weaviate/weaviate/adapters/repos/db/roaringset"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector/dynamic"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector/flat"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector/noop"
 	"github.com/weaviate/weaviate/entities/additional"
 	"github.com/weaviate/weaviate/entities/aggregation"
 	"github.com/weaviate/weaviate/entities/backup"
-	"github.com/weaviate/weaviate/entities/cyclemanager"
 	"github.com/weaviate/weaviate/entities/filters"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/entities/multi"
@@ -53,11 +43,6 @@ import (
 	"github.com/weaviate/weaviate/entities/searchparams"
 	"github.com/weaviate/weaviate/entities/storagestate"
 	"github.com/weaviate/weaviate/entities/storobj"
-	"github.com/weaviate/weaviate/entities/vectorindex"
-	"github.com/weaviate/weaviate/entities/vectorindex/common"
-	dynamicent "github.com/weaviate/weaviate/entities/vectorindex/dynamic"
-	flatent "github.com/weaviate/weaviate/entities/vectorindex/flat"
-	hnswent "github.com/weaviate/weaviate/entities/vectorindex/hnsw"
 	"github.com/weaviate/weaviate/usecases/modules"
 	"github.com/weaviate/weaviate/usecases/monitoring"
 	"github.com/weaviate/weaviate/usecases/objects"
@@ -227,413 +212,6 @@ type Shard struct {
 	shutdownLock *sync.RWMutex
 }
 
-func NewShard(ctx context.Context, promMetrics *monitoring.PrometheusMetrics,
-	shardName string, index *Index, class *models.Class, jobQueueCh chan job,
-	indexCheckpoints *indexcheckpoint.Checkpoints,
-) (_ *Shard, err error) {
-	before := time.Now()
-
-	index.logger.WithFields(logrus.Fields{
-		"action": "init_shard",
-		"shard":  shardName,
-		"index":  index.ID(),
-	}).Debugf("initializing shard %q", shardName)
-
-	s := &Shard{
-		index:       index,
-		name:        shardName,
-		promMetrics: promMetrics,
-		metrics: NewMetrics(index.logger, promMetrics,
-			string(index.Config.ClassName), shardName),
-		slowQueryReporter:     helpers.NewSlowQueryReporterFromEnv(index.logger),
-		stopDimensionTracking: make(chan struct{}),
-		replicationMap:        pendingReplicaTasks{Tasks: make(map[string]replicaTask, 32)},
-		centralJobQueue:       jobQueueCh,
-		indexCheckpoints:      indexCheckpoints,
-
-		shut:         false,
-		shutdownLock: new(sync.RWMutex),
-
-		status: storagestate.StatusLoading,
-	}
-
-	defer func() {
-		p := recover()
-		if p != nil {
-			err = fmt.Errorf("unexpected error initializing shard %q of index %q: %v", shardName, index.ID(), p)
-			index.logger.WithError(err).WithFields(logrus.Fields{
-				"index": index.ID(),
-				"shard": shardName,
-			}).Error("panic during shard initialization")
-			debug.PrintStack()
-		}
-
-		if err != nil {
-			// Initializing a shard should normally not fail. If it does, this could
-			// mean that this setup requires further attention, e.g. to manually fix
-			// a data corruption. This makes it a prime use case for sentry:
-			entsentry.CaptureException(err)
-			// spawn a new context as we cannot guarantee that the init context is
-			// still valid, but we want to make sure that we have enough time to clean
-			// up the partial init
-			ctx, cancel := context.WithTimeout(context.Background(), 120*time.Second)
-			defer cancel()
-			s.index.logger.WithFields(logrus.Fields{
-				"action":   "new_shard",
-				"duration": 120 * time.Second,
-			}).Debug("context.WithTimeout")
-
-			s.cleanupPartialInit(ctx)
-		}
-	}()
-
-	s.activityTracker.Store(1) // initial state
-	s.initCycleCallbacks()
-
-	s.docIdLock = make([]sync.Mutex, IdLockPoolSize)
-
-	defer s.metrics.ShardStartup(before)
-
-	_, err = os.Stat(s.path())
-	exists := false
-	if err == nil {
-		exists = true
-	}
-
-	if err := os.MkdirAll(s.path(), os.ModePerm); err != nil {
-		return nil, err
-	}
-
-	if err := s.initNonVector(ctx, class); err != nil {
-		return nil, errors.Wrapf(err, "init shard %q", s.ID())
-	}
-
-	if s.hasTargetVectors() {
-		if err := s.initTargetVectors(ctx); err != nil {
-			return nil, err
-		}
-		if err := s.initTargetQueues(); err != nil {
-			return nil, err
-		}
-	} else {
-		if err := s.initLegacyVector(ctx); err != nil {
-			return nil, err
-		}
-		if err := s.initLegacyQueue(); err != nil {
-			return nil, err
-		}
-	}
-
-	s.initDimensionTracking()
-
-	if asyncEnabled() {
-		f := func() {
-			// preload unindexed objects in the background
-			if s.hasTargetVectors() {
-				for targetVector, queue := range s.queues {
-					err := queue.PreloadShard(s)
-					if err != nil {
-						queue.Logger.WithError(err).Errorf("preload shard for target vector: %s", targetVector)
-					}
-				}
-			} else {
-				err := s.queue.PreloadShard(s)
-				if err != nil {
-					s.queue.Logger.WithError(err).Error("preload shard")
-				}
-			}
-		}
-		enterrors.GoWrapper(f, s.index.logger)
-	}
-	s.NotifyReady()
-
-	if exists {
-		s.index.logger.Printf("Completed loading shard %s in %s", s.ID(), time.Since(before))
-	} else {
-		s.index.logger.Printf("Created shard %s in %s", s.ID(), time.Since(before))
-	}
-	return s, nil
-}
-
-func (s *Shard) hasTargetVectors() bool {
-	return hasTargetVectors(s.index.vectorIndexUserConfig, s.index.vectorIndexUserConfigs)
-}
-
-// target vectors and legacy vector are (supposed to be) exclusive
-// method allows to distinguish which of them is configured for the class
-func hasTargetVectors(cfg schemaConfig.VectorIndexConfig, targetCfgs map[string]schemaConfig.VectorIndexConfig) bool {
-	return len(targetCfgs) != 0
-}
-
-func (s *Shard) initTargetVectors(ctx context.Context) error {
-	s.vectorIndexes = make(map[string]VectorIndex)
-	for targetVector, vectorIndexConfig := range s.index.vectorIndexUserConfigs {
-		vectorIndex, err := s.initVectorIndex(ctx, targetVector, vectorIndexConfig)
-		if err != nil {
-			return fmt.Errorf("cannot create vector index for %q: %w", targetVector, err)
-		}
-		s.vectorIndexes[targetVector] = vectorIndex
-	}
-	return nil
-}
-
-func (s *Shard) initTargetQueues() error {
-	s.queues = make(map[string]*IndexQueue)
-	for targetVector, vectorIndex := range s.vectorIndexes {
-		queue, err := NewIndexQueue(s.index.Config.ClassName.String(), s.ID(), targetVector, s, vectorIndex, s.centralJobQueue,
-			s.indexCheckpoints, IndexQueueOptions{Logger: s.index.logger}, s.promMetrics)
-		if err != nil {
-			return fmt.Errorf("cannot create index queue for %q: %w", targetVector, err)
-		}
-		s.queues[targetVector] = queue
-	}
-	return nil
-}
-
-func (s *Shard) initLegacyVector(ctx context.Context) error {
-	vectorindex, err := s.initVectorIndex(ctx, "", s.index.vectorIndexUserConfig)
-	if err != nil {
-		return err
-	}
-	s.vectorIndex = vectorindex
-	return nil
-}
-
-func (s *Shard) initLegacyQueue() error {
-	queue, err := NewIndexQueue(s.index.Config.ClassName.String(), s.ID(), "", s, s.vectorIndex, s.centralJobQueue,
-		s.indexCheckpoints, IndexQueueOptions{Logger: s.index.logger}, s.promMetrics)
-	if err != nil {
-		return err
-	}
-	s.queue = queue
-	return nil
-}
-
-func (s *Shard) initVectorIndex(ctx context.Context,
-	targetVector string, vectorIndexUserConfig schemaConfig.VectorIndexConfig,
-) (VectorIndex, error) {
-	var distProv distancer.Provider
-
-	switch vectorIndexUserConfig.DistanceName() {
-	case "", common.DistanceCosine:
-		distProv = distancer.NewCosineDistanceProvider()
-	case common.DistanceDot:
-		distProv = distancer.NewDotProductProvider()
-	case common.DistanceL2Squared:
-		distProv = distancer.NewL2SquaredProvider()
-	case common.DistanceManhattan:
-		distProv = distancer.NewManhattanProvider()
-	case common.DistanceHamming:
-		distProv = distancer.NewHammingProvider()
-	default:
-		return nil, fmt.Errorf("init vector index: %w",
-			errors.Errorf("unrecognized distance metric %q,"+
-				"choose one of [\"cosine\", \"dot\", \"l2-squared\", \"manhattan\",\"hamming\"]", vectorIndexUserConfig.DistanceName()))
-	}
-
-	var vectorIndex VectorIndex
-
-	switch vectorIndexUserConfig.IndexType() {
-	case vectorindex.VectorIndexTypeHNSW:
-		hnswUserConfig, ok := vectorIndexUserConfig.(hnswent.UserConfig)
-		if !ok {
-			return nil, errors.Errorf("hnsw vector index: config is not hnsw.UserConfig: %T",
-				vectorIndexUserConfig)
-		}
-
-		if hnswUserConfig.Skip {
-			vectorIndex = noop.NewIndex()
-		} else {
-			// starts vector cycles if vector is configured
-			s.index.cycleCallbacks.vectorCommitLoggerCycle.Start()
-			s.index.cycleCallbacks.vectorTombstoneCleanupCycle.Start()
-
-			// a shard can actually have multiple vector indexes:
-			// - the main index, which is used for all normal object vectors
-			// - a geo property index for each geo prop in the schema
-			//
-			// here we label the main vector index as such.
-			vecIdxID := s.vectorIndexID(targetVector)
-
-			vi, err := hnsw.New(hnsw.Config{
-				Logger:               s.index.logger,
-				RootPath:             s.path(),
-				ID:                   vecIdxID,
-				ShardName:            s.name,
-				ClassName:            s.index.Config.ClassName.String(),
-				PrometheusMetrics:    s.promMetrics,
-				VectorForIDThunk:     hnsw.NewVectorForIDThunk(targetVector, s.vectorByIndexID),
-				TempVectorForIDThunk: hnsw.NewTempVectorForIDThunk(targetVector, s.readVectorByIndexIDIntoSlice),
-				DistanceProvider:     distProv,
-				MakeCommitLoggerThunk: func() (hnsw.CommitLogger, error) {
-					return hnsw.NewCommitLogger(s.path(), vecIdxID,
-						s.index.logger, s.cycleCallbacks.vectorCommitLoggerCallbacks,
-						hnsw.WithAllocChecker(s.index.allocChecker),
-						hnsw.WithCommitlogThresholdForCombining(s.index.Config.HNSWMaxLogSize),
-						// consistent with previous logic where the individual limit is 1/5 of the combined limit
-						hnsw.WithCommitlogThreshold(s.index.Config.HNSWMaxLogSize/5),
-					)
-				},
-				AllocChecker:        s.index.allocChecker,
-				WaitForCachePrefill: s.index.Config.HNSWWaitForCachePrefill,
-			}, hnswUserConfig, s.cycleCallbacks.vectorTombstoneCleanupCallbacks,
-				s.cycleCallbacks.compactionCallbacks, s.cycleCallbacks.flushCallbacks, s.store)
-			if err != nil {
-				return nil, errors.Wrapf(err, "init shard %q: hnsw index", s.ID())
-			}
-			vectorIndex = vi
-		}
-	case vectorindex.VectorIndexTypeFLAT:
-		flatUserConfig, ok := vectorIndexUserConfig.(flatent.UserConfig)
-		if !ok {
-			return nil, errors.Errorf("flat vector index: config is not flat.UserConfig: %T",
-				vectorIndexUserConfig)
-		}
-		s.index.cycleCallbacks.vectorCommitLoggerCycle.Start()
-
-		// a shard can actually have multiple vector indexes:
-		// - the main index, which is used for all normal object vectors
-		// - a geo property index for each geo prop in the schema
-		//
-		// here we label the main vector index as such.
-		vecIdxID := s.vectorIndexID(targetVector)
-
-		vi, err := flat.New(flat.Config{
-			ID:               vecIdxID,
-			TargetVector:     targetVector,
-			Logger:           s.index.logger,
-			DistanceProvider: distProv,
-			AllocChecker:     s.index.allocChecker,
-		}, flatUserConfig, s.store)
-		if err != nil {
-			return nil, errors.Wrapf(err, "init shard %q: flat index", s.ID())
-		}
-		vectorIndex = vi
-	case vectorindex.VectorIndexTypeDYNAMIC:
-		dynamicUserConfig, ok := vectorIndexUserConfig.(dynamicent.UserConfig)
-		if !ok {
-			return nil, errors.Errorf("dynamic vector index: config is not dynamic.UserConfig: %T",
-				vectorIndexUserConfig)
-		}
-		s.index.cycleCallbacks.vectorCommitLoggerCycle.Start()
-
-		// a shard can actually have multiple vector indexes:
-		// - the main index, which is used for all normal object vectors
-		// - a geo property index for each geo prop in the schema
-		//
-		// here we label the main vector index as such.
-		vecIdxID := s.vectorIndexID(targetVector)
-
-		vi, err := dynamic.New(dynamic.Config{
-			ID:                   vecIdxID,
-			TargetVector:         targetVector,
-			Logger:               s.index.logger,
-			DistanceProvider:     distProv,
-			RootPath:             s.path(),
-			ShardName:            s.name,
-			ClassName:            s.index.Config.ClassName.String(),
-			PrometheusMetrics:    s.promMetrics,
-			VectorForIDThunk:     hnsw.NewVectorForIDThunk(targetVector, s.vectorByIndexID),
-			TempVectorForIDThunk: hnsw.NewTempVectorForIDThunk(targetVector, s.readVectorByIndexIDIntoSlice),
-			MakeCommitLoggerThunk: func() (hnsw.CommitLogger, error) {
-				return hnsw.NewCommitLogger(s.path(), vecIdxID,
-					s.index.logger, s.cycleCallbacks.vectorCommitLoggerCallbacks)
-			},
-			TombstoneCallbacks:       s.cycleCallbacks.vectorTombstoneCleanupCallbacks,
-			ShardCompactionCallbacks: s.cycleCallbacks.compactionCallbacks,
-			ShardFlushCallbacks:      s.cycleCallbacks.flushCallbacks,
-		}, dynamicUserConfig, s.store)
-		if err != nil {
-			return nil, errors.Wrapf(err, "init shard %q: dynamic index", s.ID())
-		}
-		vectorIndex = vi
-	default:
-		return nil, fmt.Errorf("Unknown vector index type: %q. Choose one from [\"%s\", \"%s\", \"%s\"]",
-			vectorIndexUserConfig.IndexType(), vectorindex.VectorIndexTypeHNSW, vectorindex.VectorIndexTypeFLAT, vectorindex.VectorIndexTypeDYNAMIC)
-	}
-	defer vectorIndex.PostStartup()
-	return vectorIndex, nil
-}
-
-func (s *Shard) initNonVector(ctx context.Context, class *models.Class) error {
-	before := time.Now()
-	defer func() {
-		took := time.Since(before)
-		s.index.logger.WithFields(logrus.Fields{
-			"action":   "init_shard_non_vector",
-			"duration": took,
-		}).Debugf("loaded non-vector (lsm, object, inverted) in %s for shard %q", took, s.ID())
-	}()
-
-	// init the store itself synchronously
-	err := s.initLSMStore()
-	if err != nil {
-		return fmt.Errorf("init shard %q: lsm store: %w", s.ID(), err)
-	}
-
-	// the shard versioner is also dependency of some of the bucket
-	// initializations, so it also needs to happen synchronously
-	if err := s.initIndexCounterVersionerAndBitmapFactory(); err != nil {
-		return fmt.Errorf("init shard %q: %w", s.ID(), err)
-	}
-
-	// Run all other inits in parallel and use a single error group to wait for
-	// all init tasks, the wait statement is at the end of this method. No other
-	// methods should attempt to wait on this error group.
-	eg := enterrors.NewErrorGroupWrapper(s.index.logger)
-
-	eg.Go(func() error {
-		return s.initObjectBucket(ctx)
-	})
-
-	eg.Go(func() error {
-		return s.initProplenTracker()
-	})
-
-	// error group is passed, so properties can be initialized in parallel with
-	// the other initializations going on here.
-	s.initProperties(eg, class)
-
-	err = eg.Wait()
-	if err != nil {
-		// annotate error with shard id only once, all inner functions should only
-		// annotate what they do, but not repeat the shard id.
-		return fmt.Errorf("init shard %q: %w", s.ID(), err)
-	}
-
-	return nil
-}
-
-func (s *Shard) initProplenTracker() error {
-	plPath := path.Join(s.path(), "proplengths")
-	tracker, err := inverted.NewJsonShardMetaData(plPath, s.index.logger)
-	if err != nil {
-		return fmt.Errorf("init prop length tracker: %w", err)
-	}
-
-	s.propLenTracker = tracker
-	return nil
-}
-
-func (s *Shard) initIndexCounterVersionerAndBitmapFactory() error {
-	counter, err := indexcounter.New(s.path())
-	if err != nil {
-		return fmt.Errorf("init index counter: %w", err)
-	}
-	s.counter = counter
-	s.bitmapFactory = roaringset.NewBitmapFactory(s.counter.Get, s.index.logger)
-
-	dataPresent := s.counter.PreviewNext() != 0
-	versionPath := path.Join(s.path(), "version")
-	versioner, err := newShardVersioner(versionPath, dataPresent)
-	if err != nil {
-		return fmt.Errorf("init shard versioner: %w", err)
-	}
-	s.versioner = versioner
-
-	return nil
-}
-
 func (s *Shard) ID() string {
 	return shardId(s.index.ID(), s.name)
 }
@@ -659,233 +237,6 @@ func (s *Shard) uuidToIdLockPoolId(idBytes []byte) uint8 {
 	return idBytes[15] % IdLockPoolSize
 }
 
-func (s *Shard) initLSMStore() error {
-	annotatedLogger := s.index.logger.WithFields(logrus.Fields{
-		"shard": s.name,
-		"index": s.index.ID(),
-		"class": s.index.Config.ClassName,
-	})
-	var metrics *lsmkv.Metrics
-	if s.promMetrics != nil {
-		metrics = lsmkv.NewMetrics(s.promMetrics, string(s.index.Config.ClassName), s.name)
-	}
-
-	store, err := lsmkv.New(s.pathLSM(), s.path(), annotatedLogger, metrics,
-		s.cycleCallbacks.compactionCallbacks, s.cycleCallbacks.flushCallbacks)
-	if err != nil {
-		return fmt.Errorf("init lsmkv store at %s: %w", s.pathLSM(), err)
-	}
-
-	s.store = store
-
-	return nil
-}
-
-func (s *Shard) initObjectBucket(ctx context.Context) error {
-	opts := []lsmkv.BucketOption{
-		lsmkv.WithStrategy(lsmkv.StrategyReplace),
-		lsmkv.WithSecondaryIndices(1),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithKeepTombstones(true),
-		s.dynamicMemtableSizing(),
-		s.memtableDirtyConfig(),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	}
-
-	if s.metrics != nil && !s.metrics.grouped {
-		// If metrics are grouped we cannot observe the count of an individual
-		// shard's object store because there is just a single metric. We would
-		// override it. See https://github.com/weaviate/weaviate/issues/4396 for
-		// details.
-		opts = append(opts, lsmkv.WithMonitorCount())
-	}
-	err := s.store.CreateOrLoadBucket(ctx, helpers.ObjectsBucketLSM, opts...)
-	if err != nil {
-		return fmt.Errorf("create objects bucket: %w", err)
-	}
-
-	return nil
-}
-
-// IMPORTANT:
-// Be advised there exists LazyLoadShard::drop() implementation intended
-// to drop shard that was not loaded (instantiated) yet.
-// It deletes shard by performing required actions and removing entire shard directory.
-// If there is any action that needs to be performed beside files/dirs being removed
-// from shard directory, it needs to be reflected as well in LazyLoadShard::drop()
-// method to keep drop behaviour consistent.
-func (s *Shard) drop() (err error) {
-	s.metrics.DeleteShardLabels(s.index.Config.ClassName.String(), s.name)
-	s.metrics.baseMetrics.StartUnloadingShard(s.index.Config.ClassName.String())
-	s.replicationMap.clear()
-
-	if s.index.Config.TrackVectorDimensions {
-		// tracking vector dimensions goroutine only works when tracking is enabled
-		// that's why we are trying to stop it only in this case
-		s.stopDimensionTracking <- struct{}{}
-		// send 0 in when index gets dropped
-		s.clearDimensionMetrics()
-	}
-
-	ctx, cancel := context.WithTimeout(context.TODO(), 20*time.Second)
-	defer cancel()
-	s.index.logger.WithFields(logrus.Fields{
-		"action":   "drop_shard",
-		"duration": 5 * time.Second,
-	}).Debug("context.WithTimeout")
-
-	// unregister all callbacks at once, in parallel
-	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
-		s.cycleCallbacks.compactionCallbacksCtrl,
-		s.cycleCallbacks.flushCallbacksCtrl,
-		s.cycleCallbacks.vectorCombinedCallbacksCtrl,
-		s.cycleCallbacks.geoPropsCombinedCallbacksCtrl,
-	).Unregister(ctx); err != nil {
-		return err
-	}
-
-	if err = s.store.Shutdown(ctx); err != nil {
-		return errors.Wrap(err, "stop lsmkv store")
-	}
-
-	if _, err = os.Stat(s.pathLSM()); err == nil {
-		err := os.RemoveAll(s.pathLSM())
-		if err != nil {
-			return errors.Wrapf(err, "remove lsm store at %s", s.pathLSM())
-		}
-	}
-	// delete indexcount
-	err = s.counter.Drop()
-	if err != nil {
-		return errors.Wrapf(err, "remove indexcount at %s", s.path())
-	}
-
-	// delete version
-	err = s.versioner.Drop()
-	if err != nil {
-		return errors.Wrapf(err, "remove version at %s", s.path())
-	}
-
-	if s.hasTargetVectors() {
-		// TODO run in parallel?
-		for targetVector, queue := range s.queues {
-			if err = queue.Drop(); err != nil {
-				return fmt.Errorf("close queue of vector %q at %s: %w", targetVector, s.path(), err)
-			}
-		}
-		for targetVector, vectorIndex := range s.vectorIndexes {
-			if err = vectorIndex.Drop(ctx); err != nil {
-				return fmt.Errorf("remove vector index of vector %q at %s: %w", targetVector, s.path(), err)
-			}
-		}
-	} else {
-		// delete queue cursor
-		if err = s.queue.Drop(); err != nil {
-			return errors.Wrapf(err, "close queue at %s", s.path())
-		}
-		// remove vector index
-		if err = s.vectorIndex.Drop(ctx); err != nil {
-			return errors.Wrapf(err, "remove vector index at %s", s.path())
-		}
-	}
-
-	// delete property length tracker
-	err = s.GetPropertyLengthTracker().Drop()
-	if err != nil {
-		return errors.Wrapf(err, "remove prop length tracker at %s", s.path())
-	}
-
-	s.propertyIndicesLock.Lock()
-	err = s.propertyIndices.DropAll(ctx)
-	s.propertyIndicesLock.Unlock()
-	if err != nil {
-		return errors.Wrapf(err, "remove property specific indices at %s", s.path())
-	}
-
-	s.metrics.baseMetrics.FinishUnloadingShard(s.index.Config.ClassName.String())
-
-	return nil
-}
-
-func (s *Shard) addIDProperty(ctx context.Context) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	err := s.store.CreateOrLoadBucket(ctx,
-		helpers.BucketFromPropNameLSM(filters.InternalPropID),
-		s.memtableDirtyConfig(),
-		lsmkv.WithStrategy(lsmkv.StrategySetCollection),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	)
-	if err != nil {
-		return fmt.Errorf("create id property: %w", err)
-	}
-	return nil
-}
-
-func (s *Shard) addDimensionsProperty(ctx context.Context) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	// Note: this data would fit the "Set" type better, but since the "Map" type
-	// is currently optimized better, it is more efficient to use a Map here.
-	err := s.store.CreateOrLoadBucket(ctx,
-		helpers.DimensionsBucketLSM,
-		lsmkv.WithStrategy(lsmkv.StrategyMapCollection),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	)
-	if err != nil {
-		return fmt.Errorf("create dimensions tracking property: %w", err)
-	}
-
-	return nil
-}
-
-func (s *Shard) addTimestampProperties(ctx context.Context) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	if err := s.addCreationTimeUnixProperty(ctx); err != nil {
-		return fmt.Errorf("create creation time property: %w", err)
-	}
-
-	if err := s.addLastUpdateTimeUnixProperty(ctx); err != nil {
-		return fmt.Errorf("create last update time property: %w", err)
-	}
-
-	return nil
-}
-
-func (s *Shard) addCreationTimeUnixProperty(ctx context.Context) error {
-	return s.store.CreateOrLoadBucket(ctx,
-		helpers.BucketFromPropNameLSM(filters.InternalPropCreationTimeUnix),
-		s.memtableDirtyConfig(),
-		lsmkv.WithStrategy(lsmkv.StrategyRoaringSet),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	)
-}
-
-func (s *Shard) addLastUpdateTimeUnixProperty(ctx context.Context) error {
-	return s.store.CreateOrLoadBucket(ctx,
-		helpers.BucketFromPropNameLSM(filters.InternalPropLastUpdateTimeUnix),
-		s.memtableDirtyConfig(),
-		lsmkv.WithStrategy(lsmkv.StrategyRoaringSet),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	)
-}
-
 func (s *Shard) memtableDirtyConfig() lsmkv.BucketOption {
 	return lsmkv.WithDirtyThreshold(
 		time.Duration(s.index.Config.MemtablesFlushDirtyAfter) * time.Second)
@@ -897,131 +248,6 @@ func (s *Shard) dynamicMemtableSizing() lsmkv.BucketOption {
 		s.index.Config.MemtablesMaxSizeMB,
 		s.index.Config.MemtablesMinActiveSeconds,
 		s.index.Config.MemtablesMaxActiveSeconds,
-	)
-}
-
-func (s *Shard) initPropertyBuckets(ctx context.Context, eg *enterrors.ErrorGroupWrapper, props ...*models.Property) {
-	for _, prop := range props {
-		if !inverted.HasInvertedIndex(prop) {
-			continue
-		}
-
-		propCopy := *prop // prevent loop variable capture
-
-		eg.Go(func() error {
-			if err := s.createPropertyValueIndex(ctx, &propCopy); err != nil {
-				return fmt.Errorf("init prop %q: value index: %w", propCopy.Name, err)
-			}
-			return nil
-		})
-
-		if s.index.invertedIndexConfig.IndexNullState {
-			eg.Go(func() error {
-				if err := s.createPropertyNullIndex(ctx, &propCopy); err != nil {
-					return fmt.Errorf("init prop %q: null index: %w", prop.Name, err)
-				}
-				return nil
-			})
-		}
-
-		if s.index.invertedIndexConfig.IndexPropertyLength {
-			eg.Go(func() error {
-				if err := s.createPropertyLengthIndex(ctx, &propCopy); err != nil {
-					return fmt.Errorf("init prop %q: length index: %w", prop.Name, err)
-				}
-				return nil
-			})
-		}
-
-	}
-}
-
-func (s *Shard) createPropertyValueIndex(ctx context.Context, prop *models.Property) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	bucketOpts := []lsmkv.BucketOption{
-		s.memtableDirtyConfig(),
-		s.dynamicMemtableSizing(),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	}
-
-	if inverted.HasFilterableIndex(prop) {
-		if dt, _ := schema.AsPrimitive(prop.DataType); dt == schema.DataTypeGeoCoordinates {
-			return s.initGeoProp(prop)
-		}
-
-		if schema.IsRefDataType(prop.DataType) {
-			if err := s.store.CreateOrLoadBucket(ctx,
-				helpers.BucketFromPropNameMetaCountLSM(prop.Name),
-				append(bucketOpts, lsmkv.WithStrategy(lsmkv.StrategyRoaringSet))...,
-			); err != nil {
-				return err
-			}
-		}
-
-		if err := s.store.CreateOrLoadBucket(ctx,
-			helpers.BucketFromPropNameLSM(prop.Name),
-			append(bucketOpts, lsmkv.WithStrategy(lsmkv.StrategyRoaringSet))...,
-		); err != nil {
-			return err
-		}
-	}
-
-	if inverted.HasSearchableIndex(prop) {
-		searchableBucketOpts := append(bucketOpts,
-			lsmkv.WithStrategy(lsmkv.StrategyMapCollection), lsmkv.WithPread(s.index.Config.AvoidMMap))
-		if s.versioner.Version() < 2 {
-			searchableBucketOpts = append(searchableBucketOpts, lsmkv.WithLegacyMapSorting())
-		}
-
-		if err := s.store.CreateOrLoadBucket(ctx,
-			helpers.BucketSearchableFromPropNameLSM(prop.Name),
-			searchableBucketOpts...,
-		); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (s *Shard) createPropertyLengthIndex(ctx context.Context, prop *models.Property) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	// some datatypes are not added to the inverted index, so we can skip them here
-	switch schema.DataType(prop.DataType[0]) {
-	case schema.DataTypeGeoCoordinates, schema.DataTypePhoneNumber, schema.DataTypeBlob, schema.DataTypeInt,
-		schema.DataTypeNumber, schema.DataTypeBoolean, schema.DataTypeDate:
-		return nil
-	default:
-	}
-
-	return s.store.CreateOrLoadBucket(ctx,
-		helpers.BucketFromPropNameLengthLSM(prop.Name),
-		lsmkv.WithStrategy(lsmkv.StrategyRoaringSet),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
-	)
-}
-
-func (s *Shard) createPropertyNullIndex(ctx context.Context, prop *models.Property) error {
-	if s.isReadOnly() {
-		return storagestate.ErrStatusReadOnly
-	}
-
-	return s.store.CreateOrLoadBucket(ctx,
-		helpers.BucketFromPropNameNullLSM(prop.Name),
-		lsmkv.WithStrategy(lsmkv.StrategyRoaringSet),
-		lsmkv.WithPread(s.index.Config.AvoidMMap),
-		lsmkv.WithAllocChecker(s.index.allocChecker),
-		lsmkv.WithMaxSegmentSize(s.index.Config.MaxSegmentSize),
 	)
 }
 
@@ -1064,196 +290,6 @@ func (s *Shard) UpdateVectorIndexConfigs(ctx context.Context, updated map[string
 	enterrors.GoWrapper(f, s.index.logger)
 
 	return err
-}
-
-/*
-
-	batch
-		shut
-		false
-			in_use ++
-			defer in_use --
-		true
-			fail request
-
-	shutdown
-		loop + time:
-		if shut == true
-			fail request
-		in_use == 0 && shut == false
-			shut = true
-
-*/
-// Shutdown needs to be idempotent, so it can also deal with a partial
-// initialization. In some parts, it relies on the underlying structs to have
-// idempotent Shutdown methods. In other parts, it explicitly checks if a
-// component was initialized. If not, it turns it into a noop to prevent
-// blocking.
-func (s *Shard) Shutdown(ctx context.Context) (err error) {
-	if err = s.waitForShutdown(ctx); err != nil {
-		return
-	}
-
-	if err = s.GetPropertyLengthTracker().Close(); err != nil {
-		return errors.Wrap(err, "close prop length tracker")
-	}
-
-	// unregister all callbacks at once, in parallel
-	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
-		s.cycleCallbacks.compactionCallbacksCtrl,
-		s.cycleCallbacks.flushCallbacksCtrl,
-		s.cycleCallbacks.vectorCombinedCallbacksCtrl,
-		s.cycleCallbacks.geoPropsCombinedCallbacksCtrl,
-	).Unregister(ctx); err != nil {
-		return err
-	}
-
-	if s.hasTargetVectors() {
-		// TODO run in parallel?
-		for targetVector, queue := range s.queues {
-			if err = queue.Close(); err != nil {
-				return fmt.Errorf("shut down vector index queue of vector %q: %w", targetVector, err)
-			}
-		}
-		for targetVector, vectorIndex := range s.vectorIndexes {
-			if vectorIndex == nil {
-				// a nil-vector index during shutdown would indicate that the shard was not
-				// fully initialized, the vector index shutdown becomes a no-op
-				continue
-			}
-
-			if err = vectorIndex.Flush(); err != nil {
-				return fmt.Errorf("flush vector index commitlog of vector %q: %w", targetVector, err)
-			}
-			if err = vectorIndex.Shutdown(ctx); err != nil {
-				return fmt.Errorf("shut down vector index of vector %q: %w", targetVector, err)
-			}
-		}
-	} else {
-		if err = s.queue.Close(); err != nil {
-			return errors.Wrap(err, "shut down vector index queue")
-		}
-		if s.vectorIndex != nil {
-			// a nil-vector index during shutdown would indicate that the shard was not
-			// fully initialized, the vector index shutdown becomes a no-op
-
-			// to ensure that all commitlog entries are written to disk.
-			// otherwise in some cases the tombstone cleanup process'
-			// 'RemoveTombstone' entry is not picked up on restarts
-			// resulting in perpetually attempting to remove a tombstone
-			// which doesn't actually exist anymore
-			if err = s.vectorIndex.Flush(); err != nil {
-				return errors.Wrap(err, "flush vector index commitlog")
-			}
-			if err = s.vectorIndex.Shutdown(ctx); err != nil {
-				return errors.Wrap(err, "shut down vector index")
-			}
-		}
-	}
-
-	// unregister all callbacks at once, in parallel
-	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
-		s.cycleCallbacks.compactionCallbacksCtrl,
-		s.cycleCallbacks.flushCallbacksCtrl,
-		s.cycleCallbacks.vectorCombinedCallbacksCtrl,
-		s.cycleCallbacks.geoPropsCombinedCallbacksCtrl,
-	).Unregister(ctx); err != nil {
-		return err
-	}
-
-	if s.store != nil {
-		// store would be nil if loading the objects bucket failed, as we would
-		// only return the store on success from s.initLSMStore()
-		if err = s.store.Shutdown(ctx); err != nil {
-			return errors.Wrap(err, "stop lsmkv store")
-		}
-	}
-
-	if s.dimensionTrackingInitialized.Load() {
-		// tracking vector dimensions goroutine only works when tracking is enabled
-		// _and_ when initialization completed, that's why we are trying to stop it
-		// only in this case
-		s.stopDimensionTracking <- struct{}{}
-	}
-
-	return nil
-}
-
-// cleanupPartialInit is called when the shard was only partially initialized.
-// Internally it just uses [Shutdown], but also adds some logging.
-func (s *Shard) cleanupPartialInit(ctx context.Context) {
-	log := s.index.logger.WithField("action", "cleanup_partial_initialization")
-	if err := s.Shutdown(ctx); err != nil {
-		log.WithError(err).Error("failed to shutdown store")
-	}
-
-	log.Debug("successfully cleaned up partially initialized shard")
-}
-
-func (s *Shard) preventShutdown() (release func(), err error) {
-	s.shutdownLock.RLock()
-	defer s.shutdownLock.RUnlock()
-
-	if s.shut {
-		return func() {}, errAlreadyShutdown
-	}
-
-	s.inUseCounter.Add(1)
-	return func() { s.inUseCounter.Add(-1) }, nil
-}
-
-func (s *Shard) waitForShutdown(ctx context.Context) error {
-	checkInterval := 50 * time.Millisecond
-	timeout := 30 * time.Second
-
-	ctx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	if eligible, err := s.checkEligibleForShutdown(); err != nil {
-		return err
-	} else if !eligible {
-		ticker := time.NewTicker(checkInterval)
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-ctx.Done():
-				return fmt.Errorf("Shard::proceedWithShutdown: %w", ctx.Err())
-			case <-ticker.C:
-				if eligible, err := s.checkEligibleForShutdown(); err != nil {
-					return err
-				} else if eligible {
-					return nil
-				}
-			}
-		}
-	}
-	return nil
-}
-
-// checks whether shutdown can be executed
-// (shard is not in use at the moment)
-func (s *Shard) checkEligibleForShutdown() (eligible bool, err error) {
-	s.shutdownLock.Lock()
-	defer s.shutdownLock.Unlock()
-
-	if s.shut {
-		return false, errAlreadyShutdown
-	}
-
-	if s.inUseCounter.Load() == 0 {
-		s.shut = true
-		return true, nil
-	}
-
-	return false, nil
-}
-
-func (s *Shard) NotifyReady() {
-	s.initStatus()
-	s.index.logger.
-		WithField("action", "startup").
-		Debugf("shard=%s is ready", s.name)
 }
 
 // ObjectCount returns the exact count at any moment
