+++ adapters/repos/db/shard_shutdown.go
@@ -1,0 +1,182 @@
+package db
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/pkg/errors"
+	"github.com/weaviate/weaviate/entities/cyclemanager"
+)
+
+/*
+
+	batch
+		shut
+		false
+			in_use ++
+			defer in_use --
+		true
+			fail request
+
+	shutdown
+		loop + time:
+		if shut == true
+			fail request
+		in_use == 0 && shut == false
+			shut = true
+
+*/
+// Shutdown needs to be idempotent, so it can also deal with a partial
+// initialization. In some parts, it relies on the underlying structs to have
+// idempotent Shutdown methods. In other parts, it explicitly checks if a
+// component was initialized. If not, it turns it into a noop to prevent
+// blocking.
+func (s *Shard) Shutdown(ctx context.Context) (err error) {
+	if err = s.waitForShutdown(ctx); err != nil {
+		return
+	}
+
+	if err = s.GetPropertyLengthTracker().Close(); err != nil {
+		return errors.Wrap(err, "close prop length tracker")
+	}
+
+	// unregister all callbacks at once, in parallel
+	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
+		s.cycleCallbacks.compactionCallbacksCtrl,
+		s.cycleCallbacks.flushCallbacksCtrl,
+		s.cycleCallbacks.vectorCombinedCallbacksCtrl,
+		s.cycleCallbacks.geoPropsCombinedCallbacksCtrl,
+	).Unregister(ctx); err != nil {
+		return err
+	}
+
+	if s.hasTargetVectors() {
+		// TODO run in parallel?
+		for targetVector, queue := range s.queues {
+			if err = queue.Close(); err != nil {
+				return fmt.Errorf("shut down vector index queue of vector %q: %w", targetVector, err)
+			}
+		}
+		for targetVector, vectorIndex := range s.vectorIndexes {
+			if vectorIndex == nil {
+				// a nil-vector index during shutdown would indicate that the shard was not
+				// fully initialized, the vector index shutdown becomes a no-op
+				continue
+			}
+
+			if err = vectorIndex.Flush(); err != nil {
+				return fmt.Errorf("flush vector index commitlog of vector %q: %w", targetVector, err)
+			}
+			if err = vectorIndex.Shutdown(ctx); err != nil {
+				return fmt.Errorf("shut down vector index of vector %q: %w", targetVector, err)
+			}
+		}
+	} else {
+		if err = s.queue.Close(); err != nil {
+			return errors.Wrap(err, "shut down vector index queue")
+		}
+		if s.vectorIndex != nil {
+			// a nil-vector index during shutdown would indicate that the shard was not
+			// fully initialized, the vector index shutdown becomes a no-op
+
+			// to ensure that all commitlog entries are written to disk.
+			// otherwise in some cases the tombstone cleanup process'
+			// 'RemoveTombstone' entry is not picked up on restarts
+			// resulting in perpetually attempting to remove a tombstone
+			// which doesn't actually exist anymore
+			if err = s.vectorIndex.Flush(); err != nil {
+				return errors.Wrap(err, "flush vector index commitlog")
+			}
+			if err = s.vectorIndex.Shutdown(ctx); err != nil {
+				return errors.Wrap(err, "shut down vector index")
+			}
+		}
+	}
+
+	// unregister all callbacks at once, in parallel
+	if err = cyclemanager.NewCombinedCallbackCtrl(0, s.index.logger,
+		s.cycleCallbacks.compactionCallbacksCtrl,
+		s.cycleCallbacks.flushCallbacksCtrl,
+		s.cycleCallbacks.vectorCombinedCallbacksCtrl,
+		s.cycleCallbacks.geoPropsCombinedCallbacksCtrl,
+	).Unregister(ctx); err != nil {
+		return err
+	}
+
+	if s.store != nil {
+		// store would be nil if loading the objects bucket failed, as we would
+		// only return the store on success from s.initLSMStore()
+		if err = s.store.Shutdown(ctx); err != nil {
+			return errors.Wrap(err, "stop lsmkv store")
+		}
+	}
+
+	if s.dimensionTrackingInitialized.Load() {
+		// tracking vector dimensions goroutine only works when tracking is enabled
+		// _and_ when initialization completed, that's why we are trying to stop it
+		// only in this case
+		s.stopDimensionTracking <- struct{}{}
+	}
+
+	return nil
+}
+
+func (s *Shard) preventShutdown() (release func(), err error) {
+	s.shutdownLock.RLock()
+	defer s.shutdownLock.RUnlock()
+
+	if s.shut {
+		return func() {}, errAlreadyShutdown
+	}
+
+	s.inUseCounter.Add(1)
+	return func() { s.inUseCounter.Add(-1) }, nil
+}
+
+func (s *Shard) waitForShutdown(ctx context.Context) error {
+	checkInterval := 50 * time.Millisecond
+	timeout := 30 * time.Second
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	if eligible, err := s.checkEligibleForShutdown(); err != nil {
+		return err
+	} else if !eligible {
+		ticker := time.NewTicker(checkInterval)
+		defer ticker.Stop()
+
+		for {
+			select {
+			case <-ctx.Done():
+				return fmt.Errorf("Shard::proceedWithShutdown: %w", ctx.Err())
+			case <-ticker.C:
+				if eligible, err := s.checkEligibleForShutdown(); err != nil {
+					return err
+				} else if eligible {
+					return nil
+				}
+			}
+		}
+	}
+	return nil
+}
+
+// checks whether shutdown can be executed
+// (shard is not in use at the moment)
+func (s *Shard) checkEligibleForShutdown() (eligible bool, err error) {
+	s.shutdownLock.Lock()
+	defer s.shutdownLock.Unlock()
+
+	if s.shut {
+		return false, errAlreadyShutdown
+	}
+
+	if s.inUseCounter.Load() == 0 {
+		s.shut = true
+		return true, nil
+	}
+
+	return false, nil
+}
